Rename inject to scripts
