+++ internal/scripts/go.sh
