+++ internal/scripts/golangci-lint.sh
